:recycle: Organize construction
