+++ src/components/header.tsx
@@ -14,7 +14,7 @@ import { useDisclosure } from '@mantine/hooks';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { FC } from 'react';
-import { ColorSchemaToggle } from '@/components/parts/colorSchemaToggle';
+import { ColorSchemaToggle } from '@/components/molecules/colorSchemaToggle';
 import { siteMetadata } from '@/config/siteMetadata';
 
 const HEADER_HEIGHT = 60;
