+++ src/components/molecules/articleCard.tsx
@@ -1,8 +1,8 @@
 import { createStyles, Card, Text, Group, useMantineColorScheme } from '@mantine/core';
 import Link from 'next/link';
 import React, { FC } from 'react';
-import { CategoryBadge } from '@/components/category/badge';
-import { MicroCMSPicture } from '@/components/parts/microCmsPicture';
+import { CategoryBadge } from '@/components/molecules/category/badge';
+import { MicroCMSPicture } from '@/components/molecules/microCmsPicture';
 import { FormattedDate } from '@/lib/date';
 import { Category } from '@/types/category';
